+++ packages/core/src/index.ts
@@ -1,10 +1,16 @@
 export * from './Accordion'
 export * from './AlertDialog'
 export * from './AspectRatio'
+export * from './Autocomplete'
 export * from './Avatar'
 export * from './Calendar'
 export * from './Checkbox'
 export * from './Collapsible'
+export * from './ColorArea'
+export * from './ColorField'
+export * from './ColorSlider'
+export * from './ColorSwatch'
+export * from './ColorSwatchPicker'
 export * from './Combobox'
 // utilities
 export * from './ConfigProvider'
@@ -22,6 +28,8 @@ export * from './HoverCard'
 export * from './Label'
 export * from './Listbox'
 export * from './Menubar'
+export * from './MonthPicker'
+export * from './MonthRangePicker'
 export * from './NavigationMenu'
 export * from './NumberField'
 export * from './Pagination'
@@ -56,6 +64,36 @@ export {
   useStateMachine,
   withDefault,
 } from './shared'
+// Color utilities
+export {
+  type Color,
+  type ColorChannel,
+  type ColorFormat,
+  type ColorSpace,
+  colorToHex,
+  colorToHsb,
+  colorToHsl,
+  colorToRgb,
+  colorToString,
+  convertToHsb,
+  convertToHsl,
+  convertToRgb,
+  getAreaBackgroundStyle,
+  getAreaGradient,
+  getChannelName,
+  getChannelRange,
+  getChannelValue,
+  getSliderBackgroundStyle,
+  getSliderGradient,
+  type HSBColor,
+  type HSLColor,
+  isValidColor,
+  normalizeColor,
+  parseColor,
+  type RGBColor,
+  setChannelValue,
+  setChannelValues,
+} from './shared/color'
 export {
   type AcceptableValue,
   type GenericComponentInstance,
@@ -68,6 +106,7 @@ export * from './Switch'
 export * from './Tabs'
 export * from './TagsInput'
 export * from './TimeField'
+export * from './TimeRangeField'
 
 export * from './Toast'
 export * from './Toggle'
@@ -77,6 +116,8 @@ export * from './Tooltip'
 export * from './Tree'
 export * from './Viewport'
 export { VisuallyHidden, type VisuallyHiddenProps } from './VisuallyHidden'
+export * from './YearPicker'
+export * from './YearRangePicker'
 export {
   type ReferenceElement,
 } from '@floating-ui/vue'